refactor: move address parsing utilities to connections module
